Add files via upload
1.what is emmet?
Emmet is a free add-on for your text editor. 
It allows you to type shortcuts that are then expanded into full pieces of code. 
By using Emmet, developers type less, they save both on keystrokes and time


2.Difference between Library and Framework?
A framework is like the foundation upon which developers build applications for specific platforms. 
It includes reusable pieces of code written to perform common tasks and uses code provided by a developer for custom functionality

By using a library, you can control the flow of the application and call the library


3.What is CDN? WHY DO WE USE IT?
A content delivery network (CDN) is a network of interconnected servers that speeds up webpage loading for data-heavy applications. 
CDN can stand for content delivery network or content distribution network

4.Why is react known as react?
React is an open source library which enhance or give us more power to build vast web applications 
browsers have much powerful tools and APIS what we used and upscaling that React helps us to build more in easy and convenient way 

5.What is crossorigin in script tag?
A cross-origin request is a request for a resource (e.g. style sheets, iframes, images, fonts, or scripts) from another domain
to interact with other resources we use cross origin


6.What is differnce between React and React-DOM?
So basically react is a library which helps us to create an elemnts inside it 
and inside that elemnts at the end its and object what is created 

but to render or to lift or to store that object we use React-DOM which helps for manipulating DOM
because normal DOM is costly and this is easy as well it helps to runs the whole react inside it 

7.what is react.development.js and react.production.js via CDN links?
so development.js is used to build react elements and used to build your project locally on your machine
where as production is used to run your project on production mode i.e on client side or client machine

8.what is async and defer?
async runs faster as when your script is ready or loaded it will run without blocking other elements 
means without any side effects or a paticular element is loaded it will run

whereas defer will only run once all of the element present in a single page or current page is loaded by blocking elements